DSND-2420: Extract logic into dependencies for testability
