+++ src/main/java/uk/gov/companieshouse/filinghistory/api/service/FilingHistoryStatusService.java
@@ -3,16 +3,17 @@ package uk.gov.companieshouse.filinghistory.api.service;
 import java.util.Map;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
-import uk.gov.companieshouse.filinghistory.api.statusrules.FromProperties;
 import uk.gov.companieshouse.filinghistory.api.statusrules.PrefixProperties;
 import uk.gov.companieshouse.filinghistory.api.statusrules.StatusRuleProperties;
 
 public class FilingHistoryStatusService implements StatusService {
 
     private final StatusRuleProperties statusRuleProperties;
+    private final CompanyNumberStatusProcessor companyNumberStatusProcessor;
 
-    public FilingHistoryStatusService(StatusRuleProperties statusRuleProperties) {
+    public FilingHistoryStatusService(StatusRuleProperties statusRuleProperties, CompanyNumberStatusProcessor companyNumberStatusProcessor) {
         this.statusRuleProperties = statusRuleProperties;
+        this.companyNumberStatusProcessor = companyNumberStatusProcessor;
     }
 
     @Override
@@ -20,26 +21,12 @@ public class FilingHistoryStatusService implements StatusService {
         Pattern pattern = Pattern.compile("^([A-Z]{2}|R0|)(\\d+)");
         Matcher matcher = pattern.matcher(companyNumber);
 
-        final String prefix;
-        if (matcher.find()) {
-            prefix = matcher.group(1).isEmpty() ? "NORMAL" : matcher.group(1);
-        } else {
-            prefix = "INVALID_FORMAT";
-        }
+        final String prefix = companyNumberStatusProcessor.getPrefixFromRegexMatch(matcher);
 
         Map<String, PrefixProperties> filingHistory = statusRuleProperties.filingHistory();
         PrefixProperties prefixProperties = filingHistory.getOrDefault(prefix, filingHistory.get("UNKNOWN_PREFIX"));
 
-        String status = prefixProperties.status();
-        if (prefixProperties.from() != null) {
-            for (FromProperties fromProperties : prefixProperties.from()) {
-                if (fromProperties.number() > Integer.parseInt(matcher.group(2))) {
-                    break;
-                }
-                status = fromProperties.status();
-            }
-        }
-        return status;
+        return companyNumberStatusProcessor.getStatusFromPrefixProperties(prefixProperties, matcher);
     }
 }
